+++ api/controllers/book/books_test.go
@@ -1,4 +1,4 @@
-package books
+package book
 
 import (
 	"encoding/json"
